refactor: rename core crate modules to descriptive names
Closes #445
