+++ src/lib.rs
@@ -1,21 +1,12 @@
 //! # Soroban Budget Assert — Core Module
 //!
 //! This crate provides the foundational traits and types for cost measurement,
-//! budget assertion, and resource reporting. See [`module_1`] for the full
-//! trait documentation and usage examples.
-//!
-//! ## Modules
-//!
-//! - [`module_1`] — Core cost-measurement, budget-assertion, and
-//!   reporting traits and types.
-//! - [`module_21`] — Trait implementations for budget cost measurement,
-//!   estimation, and assertion.
-//! - [`module_25`] — Optimized per-operation state-tracking backends
-//!   (linear vs. hash-based) with paired benchmarks.
+//! budget assertion, and resource reporting. See the [`traits`] module for the full
+//! API documentation and usage examples.
 
-pub mod module_1;
-pub mod module_21;
-pub mod module_25;
+pub mod impls;
+pub mod state_tracking;
+pub mod traits;
 
-pub use module_1::*;
-pub use module_25::*;
+pub use state_tracking::*;
+pub use traits::*;
